concept10-agentic/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agent-orchestration-framework"
version = "0.1.0"
description = "Async-first AI agentic orchestration framework"
readme = "README.md"
requires-python = ">=3.11"
authors = [
  { name = "Concept10 Team" }
]
license = { text = "Proprietary" }
dependencies = []

[project.optional-dependencies]
core = [
  "langgraph>=0.2,<1.0",
  "langchain-core>=0.3,<1.0",
  "langchain-openai>=0.2,<1.0",
  "pydantic>=2.6,<3.0",
  "pydantic-settings>=2.0,<3.0",
  "jinja2>=3.1,<4.0",
  "fastapi>=0.115,<1.0",
  "uvicorn[standard]>=0.30,<1.0",
  "python-dotenv>=1.0,<2.0",
  "httpx>=0.27,<1.0",
  "PyYAML>=6.0,<7.0",
  "tiktoken>=0.7,<1.0",
  "redis>=5.0,<6.0",
  "structlog>=24.1,<26.0"
]
dev = [
  "pytest>=8,<9",
  "pytest-asyncio>=0.23,<1.0",
  "ruff>=0.6,<1.0",
  "mypy>=1.10,<2.0",
  "pre-commit>=3.7,<5.0"
]
observability = [
  "langsmith>=0.1,<1.0",
  "opentelemetry-api>=1.25,<2.0",
  "opentelemetry-sdk>=1.25,<2.0",
  "opentelemetry-exporter-otlp>=1.25,<2.0",
  "opentelemetry-instrumentation-fastapi>=0.46b0,<1.0"
]
governance = [
  "guardrails-ai>=0.5,<1.0",
  "presidio-analyzer>=2.2,<3.0"
]

[tool.setuptools.packages.find]
where = ["."]
include = ["agents*", "core*", "api*", "observability*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "ASYNC"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_unused_configs = true
exclude = ["dashboard/"]
